# core: more efficient nonce-update in txpool