--- a/core/tx_noncer.go
+++ b/core/tx_noncer.go
@@ -78,8 +78,8 @@ func (txn *txNoncer) setIfLower(addr common.Address, nonce uint64) {
 	txn.nonces[addr] = nonce
 }
 
-// resetAll sets the nonces for all accounts to the given map.
-func (txn *txNoncer) resetAll(all map[common.Address]uint64) {
+// setAll sets the nonces for all accounts to the given map.
+func (txn *txNoncer) setAll(all map[common.Address]uint64) {
 	txn.lock.Lock()
 	defer txn.lock.Unlock()