This PR supersedes https://github.com/ethereum/go-ethereum/pull/21164  by @WeiLoy . The original PR was based off a fork of geth but on `master` , and I was not able to push the rebased + fixed version onto a `master` branch. 

I believe that the original implementation by @WeiLoy was correct, but it had been a bit bitrotted, the PR is quite old. 
Here are the benchmarks compared to current master: 
```
benchstat before.txt after3.txt 
name                           old time/op  new time/op  delta
PoolMultiAccountBatchInsert-6   993µs ±11%   281µs ± 7%  -71.72%  (p=0.008 n=5+5)
```