@holiman 2021-01-26
Pushed a fix to make the benchmarks call the `..Sync` methods. Stats basically the same: 
```
name                           old time/op  new time/op  delta
PoolMultiAccountBatchInsert-6  1.61ms ±15%  0.35ms ± 9%  -78.14%  (p=0.008 n=5+5)
```