@karalabe 2021-11-02
Pls add the allocs to the benchmark, just as sanity