@holiman 2021-11-02
```
name                           old time/op    new time/op    delta
PoolMultiAccountBatchInsert-6     930µs ±11%     313µs ± 7%  -66.35%  (p=0.008 n=5+5)

name                           old alloc/op   new alloc/op   delta
PoolMultiAccountBatchInsert-6    19.0kB ± 9%    19.7kB ± 8%     ~     (p=0.222 n=5+5)

name                           old allocs/op  new allocs/op  delta
PoolMultiAccountBatchInsert-6      95.6 ± 5%      92.4 ± 4%     ~     (p=0.063 n=5+5)
```